refactor db seed scripts
